pyproject.toml
[project]
name = "bilibili-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bilibili-api-python>=17.1.4",
    "curl-cffi>=0.11.1",
    "loguru>=0.7.3",
    "mcp[cli]>=1.9.0",
    "pydantic>=2.11.4",
]

[[tool.uv.index]]
url = "https://mirrors.nju.edu.cn/pypi/web/simple"
default = true

[tool.ruff]
target-version = "py312"
line-length = 200

[tool.ruff.format]
indent-style = "tab"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N"]
ignore = ["E501"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false
